+++ plant-project/main/modules/screen.h
@@ -8,4 +8,4 @@ void update_screen_data(uint8_t value, char format[], uint8_t icon[288]);
 void menu_air_temperature(float * temperatur);
 void menu_air_humidity(float * humidity);
 void menu_soil_temperature(float * soil_temperatur);
-void menu_soil_moisture(uint8_t * soil_moisture);
+void menu_soil_moisture(uint16_t * soil_moisture);
